Ports missing restify testing to versioned tests and removes integration test.
restify doesn't support Node 18. Instead of skipping, ported to versioned testing which is better for instrumentation as it tests functionality against multiple Node and restify version permutations.
